update: db to core.storage
