--- src/main/scala/io/lunes/db/VersionedStorage.scala
@@ -1,23 +1,0 @@
-package io.lunes.db
-
-/**
-  *
-  */
-trait VersionedStorage {
-  this: PropertiesStorage =>
-
-  private val VersionProperty = "version"
-
-  protected val Version: Int
-
-  /**
-    *
-    * @return
-    */
-  def isVersionValid: Boolean = getIntProperty(VersionProperty) match {
-    case None =>
-      putIntProperty(VersionProperty, Version, None)
-      true
-    case Some(v) => v == Version
-  }
-}
